Add Id class and move classes to common fields
