+++ src/main/java/gomedic/logic/commands/EditCommand.java
@@ -19,11 +19,11 @@ import gomedic.commons.core.index.Index;
 import gomedic.commons.util.CollectionUtil;
 import gomedic.logic.commands.exceptions.CommandException;
 import gomedic.model.Model;
-import gomedic.model.person.Address;
-import gomedic.model.person.Email;
-import gomedic.model.person.Name;
+import gomedic.model.commonfield.Address;
+import gomedic.model.commonfield.Email;
+import gomedic.model.commonfield.Name;
+import gomedic.model.commonfield.Phone;
 import gomedic.model.person.Person;
-import gomedic.model.person.Phone;
 import gomedic.model.tag.Tag;
 
 /**
