+++ src/main/java/gomedic/logic/parser/AddCommandParser.java
@@ -6,11 +6,11 @@ import java.util.stream.Stream;
 import gomedic.commons.core.Messages;
 import gomedic.logic.commands.AddCommand;
 import gomedic.logic.parser.exceptions.ParseException;
-import gomedic.model.person.Address;
-import gomedic.model.person.Email;
-import gomedic.model.person.Name;
+import gomedic.model.commonfield.Address;
+import gomedic.model.commonfield.Email;
+import gomedic.model.commonfield.Name;
+import gomedic.model.commonfield.Phone;
 import gomedic.model.person.Person;
-import gomedic.model.person.Phone;
 import gomedic.model.tag.Tag;
 
 /**
