+++ src/main/java/gomedic/logic/parser/ParserUtil.java
@@ -9,10 +9,10 @@ import java.util.Set;
 import gomedic.commons.core.index.Index;
 import gomedic.commons.util.StringUtil;
 import gomedic.logic.parser.exceptions.ParseException;
-import gomedic.model.person.Address;
-import gomedic.model.person.Email;
-import gomedic.model.person.Name;
-import gomedic.model.person.Phone;
+import gomedic.model.commonfield.Address;
+import gomedic.model.commonfield.Email;
+import gomedic.model.commonfield.Name;
+import gomedic.model.commonfield.Phone;
 import gomedic.model.tag.Tag;
 
 /**
