+++ src/main/java/gomedic/model/commonfield/Phone.java
@@ -1,4 +1,4 @@
-package gomedic.model.person;
+package gomedic.model.commonfield;
 
 import static java.util.Objects.requireNonNull;
 
@@ -9,8 +9,6 @@ import gomedic.commons.util.AppUtil;
  * Guarantees: immutable; is valid as declared in {@link #isValidPhone(String)}
  */
 public class Phone {
-
-
     public static final String MESSAGE_CONSTRAINTS =
             "Phone numbers should only contain numbers, and it should be at least 3 digits long";
     public static final String VALIDATION_REGEX = "\\d{3,}";
