+++ src/test/java/gomedic/logic/parser/AddCommandParserTest.java
@@ -5,11 +5,11 @@ import org.junit.jupiter.api.Test;
 import gomedic.commons.core.Messages;
 import gomedic.logic.commands.AddCommand;
 import gomedic.logic.commands.CommandTestUtil;
-import gomedic.model.person.Address;
-import gomedic.model.person.Email;
-import gomedic.model.person.Name;
+import gomedic.model.commonfield.Address;
+import gomedic.model.commonfield.Email;
+import gomedic.model.commonfield.Name;
+import gomedic.model.commonfield.Phone;
 import gomedic.model.person.Person;
-import gomedic.model.person.Phone;
 import gomedic.model.tag.Tag;
 import gomedic.testutil.PersonBuilder;
 import gomedic.testutil.TypicalPersons;
