+++ src/test/java/gomedic/logic/parser/AddressBookParserTest.java
@@ -20,8 +20,8 @@ import gomedic.logic.commands.FindCommand;
 import gomedic.logic.commands.HelpCommand;
 import gomedic.logic.commands.ListCommand;
 import gomedic.logic.parser.exceptions.ParseException;
-import gomedic.model.person.NameContainsKeywordsPredicate;
 import gomedic.model.person.Person;
+import gomedic.model.util.NameContainsKeywordsPredicate;
 import gomedic.testutil.EditPersonDescriptorBuilder;
 import gomedic.testutil.PersonBuilder;
 import gomedic.testutil.PersonUtil;
