+++ src/test/java/gomedic/logic/parser/EditCommandParserTest.java
@@ -10,10 +10,10 @@ import gomedic.commons.core.Messages;
 import gomedic.commons.core.index.Index;
 import gomedic.logic.commands.CommandTestUtil;
 import gomedic.logic.commands.EditCommand;
-import gomedic.model.person.Address;
-import gomedic.model.person.Email;
-import gomedic.model.person.Name;
-import gomedic.model.person.Phone;
+import gomedic.model.commonfield.Address;
+import gomedic.model.commonfield.Email;
+import gomedic.model.commonfield.Name;
+import gomedic.model.commonfield.Phone;
 import gomedic.model.tag.Tag;
 import gomedic.testutil.EditPersonDescriptorBuilder;
