+++ src/test/java/gomedic/logic/parser/ParserUtilTest.java
@@ -13,10 +13,10 @@ import java.util.Set;
 import org.junit.jupiter.api.Test;
 
 import gomedic.logic.parser.exceptions.ParseException;
-import gomedic.model.person.Address;
-import gomedic.model.person.Email;
-import gomedic.model.person.Name;
-import gomedic.model.person.Phone;
+import gomedic.model.commonfield.Address;
+import gomedic.model.commonfield.Email;
+import gomedic.model.commonfield.Name;
+import gomedic.model.commonfield.Phone;
 import gomedic.model.tag.Tag;
 
 public class ParserUtilTest {
