+++ src/test/java/gomedic/model/commonfield/NameContainsKeywordsPredicateTest.java
@@ -1,4 +1,4 @@
-package gomedic.model.person;
+package gomedic.model.commonfield;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -9,6 +9,7 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
+import gomedic.model.util.NameContainsKeywordsPredicate;
 import gomedic.testutil.PersonBuilder;
 
 public class NameContainsKeywordsPredicateTest {
